requirements.txt
streamlit
tensorflow
numpy
pandas
opencv-python-headless
pillow
scikit-learn